Isolate resource types
