+++ src/models/resource/is-blank-raw-resource.ts
@@ -1,0 +1,11 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+import { RawResource } from './raw-resource';
+
+export const isBlankRawResource = (value: RawResource): boolean => value.GUID === '';
